refactor: Revert changes to Makefile and test_fox.c for bonus code
The code changes in this commit revert the changes made to the Makefile and test_fox.c files for the bonus code. This is necessary because the previous changes were causing compatibility issues and impacting the functionality of the program. By reverting these changes, we ensure that the program works correctly and is compatible with the existing codebase.

+++ assignments/hw3/bonus/dmm.c
@@ -7,9 +7,9 @@ int main(int argc, char *argv[])
 {
 
     int i, j, dim;
-    int **localA;
-    int **localB;
-    int **localC;
+    double **localA;
+    double **localB;
+    double **localC;
     MPI_Init(&argc, &argv);
 
     GridInfo grid;
@@ -22,17 +22,17 @@ int main(int argc, char *argv[])
     dim = N / grid.q;
     /* allocate space for the three matrices */
 
-    localA = (int **)malloc(dim * sizeof(int *));
+    localA = (double **)malloc(dim * sizeof(double *));
 
-    localB = (int **)malloc(dim * sizeof(int *));
+    localB = (double **)malloc(dim * sizeof(double *));
 
-    localC = (int **)malloc(dim * sizeof(int *));
+    localC = (double **)malloc(dim * sizeof(double *));
 
     for (i = 0; i < dim; i++)
     {
-        *(localA + i) = (int *)malloc(dim * sizeof(int));
-        *(localB + i) = (int *)malloc(dim * sizeof(int));
-        *(localC + i) = (int *)malloc(dim * sizeof(int));
+        *(localA + i) = (double *)malloc(dim * sizeof(double));
+        *(localB + i) = (double *)malloc(dim * sizeof(double));
+        *(localC + i) = (double *)malloc(dim * sizeof(double));
     }
 
     /* Compute local matrices - Ideally the master should do this & pass it onto all the slaves */
